Ooops. Pushed to a wrong branch.
Revert "WRITING-2753 Make IntValue store an Int32 rather than a plain Int"

This reverts commit c27b032f4eddcabcfb9a008d5840d5588a61fd6b.

+++ src/Mongo/EvalCoreExpr.hs
@@ -110,8 +110,7 @@ evalCoreExpr (SetElem (i,v) a) env =
         arr <- getElements <$> evalCoreExpr a env
         val <- evalCoreExpr v env
         index <- evalCoreExpr i env
-        let updatedArr = case (take (fromIntegral index::Int) (arr ++ repeat NullValue),
-                               drop (fromIntegral index::Int) arr) of
+        let updatedArr = case (take index (arr ++ repeat NullValue), drop index arr) of
                             (x,y:ys) -> x ++ (val:ys)
                             (x,[]) -> x ++ [val]
         return Array { getElements = updatedArr }
